Crate objc2_core_image

Bindings to the CoreImage framework

See Apple’s docs and the general docs on framework crates for more information.

Structs

CIColor

The Core Image class that defines a color object.

CIImage

Apple’s documentation

Statics

kCIFormatA8⚠

Apple’s documentation

kCIFormatA16⚠

Apple’s documentation

kCIFormatABGR8⚠

Apple’s documentation

kCIFormatARGB8⚠

Apple’s documentation

kCIFormatAf⚠

Apple’s documentation

kCIFormatAh⚠

Apple’s documentation

kCIFormatBGRA8⚠

Apple’s documentation

kCIFormatL8⚠

Apple’s documentation

kCIFormatL16⚠

Apple’s documentation

kCIFormatLA8⚠

Apple’s documentation

kCIFormatLA16⚠

Apple’s documentation

kCIFormatLAf⚠

Apple’s documentation

kCIFormatLAh⚠

Apple’s documentation

kCIFormatLf⚠

Apple’s documentation

kCIFormatLh⚠

Apple’s documentation

kCIFormatR8⚠

Apple’s documentation

kCIFormatR16⚠

Apple’s documentation

kCIFormatRG8⚠

Apple’s documentation

kCIFormatRG16⚠

Apple’s documentation

kCIFormatRGB10⚠

Apple’s documentation

kCIFormatRGBA8⚠

Apple’s documentation

kCIFormatRGBA16⚠

Apple’s documentation

kCIFormatRGBAf⚠

Apple’s documentation

kCIFormatRGBAh⚠

Apple’s documentation

kCIFormatRGBX8⚠

Apple’s documentation

kCIFormatRGBX16⚠

Apple’s documentation

kCIFormatRGBXf⚠

Apple’s documentation

kCIFormatRGBXh⚠

Apple’s documentation

kCIFormatRGf⚠

Apple’s documentation

kCIFormatRGh⚠

Apple’s documentation

kCIFormatRf⚠

Apple’s documentation

kCIFormatRh⚠

Apple’s documentation

kCIImageApplyCleanAperture⚠

A Boolean value to control whether an image created with a CVPixelBuffer or an IOSurface should be cropped and offset according clean aperture attachments.

kCIImageApplyOrientationProperty⚠

Apple’s documentation

kCIImageAutoAdjustCrop⚠

Apple’s documentation

kCIImageAutoAdjustEnhance⚠

Apple’s documentation

kCIImageAutoAdjustFeatures⚠

Apple’s documentation

kCIImageAutoAdjustLevel⚠

Apple’s documentation

kCIImageAutoAdjustRedEye⚠

Apple’s documentation

kCIImageAuxiliaryDepth⚠

Apple’s documentation

kCIImageAuxiliaryDisparity⚠

Apple’s documentation

kCIImageAuxiliaryHDRGainMap⚠

Apple’s documentation

kCIImageAuxiliaryPortraitEffectsMatte⚠

Apple’s documentation

kCIImageAuxiliarySemanticSegmentationGlassesMatte⚠

Apple’s documentation

kCIImageAuxiliarySemanticSegmentationHairMatte⚠

Apple’s documentation

kCIImageAuxiliarySemanticSegmentationSkinMatte⚠

Apple’s documentation

kCIImageAuxiliarySemanticSegmentationSkyMatte⚠

Apple’s documentation

kCIImageAuxiliarySemanticSegmentationTeethMatte⚠

Apple’s documentation

kCIImageCacheImmediately⚠

Apple’s documentation

kCIImageColorSpace⚠

Apple’s documentation

kCIImageContentAverageLightLevel⚠

A value for overriding the automatic behavior of the Content Average Light Level property when creating an image.

kCIImageContentHeadroom⚠

A value for overriding the automatic behavior of the Content Headroom property when creating an image.

kCIImageExpandToHDR⚠

Apple’s documentation

kCIImageNearestSampling⚠

Apple’s documentation

kCIImageProperties⚠

Apple’s documentation

kCIImageTextureFormat⚠Deprecated

Apple’s documentation

kCIImageTextureTarget⚠Deprecated

Apple’s documentation

kCIImageToneMapHDRtoSDR⚠

Apple’s documentation

Type Aliases

CIFormat

Apple’s documentation

CIImageAutoAdjustmentOption

Apple’s documentation

CIImageOption

Apple’s documentation