objc2_core_image

Struct CIImage

pub struct CIImage { /* private fields */ }

Apple’s documentation

Implementations

impl CIImage

pub unsafe fn imageWithContentsOfURL(url: &NSURL) -> Option<Retained<CIImage>>

pub unsafe fn imageWithContentsOfURL_options( url: &NSURL, options: Option<&NSDictionary<CIImageOption, AnyObject>>, ) -> Option<Retained<CIImage>>

Safety

options generic should be of the correct type.

pub unsafe fn imageWithData(data: &NSData) -> Option<Retained<CIImage>>

pub unsafe fn imageWithData_options( data: &NSData, options: Option<&NSDictionary<CIImageOption, AnyObject>>, ) -> Option<Retained<CIImage>>

Safety

options generic should be of the correct type.

pub unsafe fn imageWithColor(color: &CIColor) -> Retained<CIImage>

pub unsafe fn emptyImage() -> Retained<CIImage>

pub unsafe fn blackImage() -> Retained<CIImage>

pub unsafe fn whiteImage() -> Retained<CIImage>

pub unsafe fn grayImage() -> Retained<CIImage>

pub unsafe fn redImage() -> Retained<CIImage>

pub unsafe fn greenImage() -> Retained<CIImage>

pub unsafe fn blueImage() -> Retained<CIImage>

pub unsafe fn cyanImage() -> Retained<CIImage>

pub unsafe fn magentaImage() -> Retained<CIImage>

pub unsafe fn yellowImage() -> Retained<CIImage>

pub unsafe fn clearImage() -> Retained<CIImage>

pub unsafe fn initWithData( this: Allocated<Self>, data: &NSData, ) -> Option<Retained<Self>>

pub unsafe fn initWithData_options( this: Allocated<Self>, data: &NSData, options: Option<&NSDictionary<CIImageOption, AnyObject>>, ) -> Option<Retained<Self>>

Safety

options generic should be of the correct type.

pub unsafe fn initWithContentsOfURL( this: Allocated<Self>, url: &NSURL, ) -> Option<Retained<Self>>

pub unsafe fn initWithContentsOfURL_options( this: Allocated<Self>, url: &NSURL, options: Option<&NSDictionary<CIImageOption, AnyObject>>, ) -> Option<Retained<Self>>

Safety

options generic should be of the correct type.

pub unsafe fn initWithColor( this: Allocated<Self>, color: &CIColor, ) -> Retained<Self>

pub unsafe fn imageByApplyingOrientation( &self, orientation: c_int, ) -> Retained<CIImage>

pub unsafe fn imageByCompositingOverImage( &self, dest: &CIImage, ) -> Retained<CIImage>

pub unsafe fn imageByClampingToExtent(&self) -> Retained<CIImage>

pub unsafe fn imageByApplyingFilter_withInputParameters( &self, filter_name: &NSString, params: &NSDictionary<NSString, AnyObject>, ) -> Retained<CIImage>

Safety

params generic should be of the correct type.

pub unsafe fn imageByApplyingFilter( &self, filter_name: &NSString, ) -> Retained<CIImage>

pub unsafe fn imageByPremultiplyingAlpha(&self) -> Retained<CIImage>

pub unsafe fn imageByUnpremultiplyingAlpha(&self) -> Retained<CIImage>

pub unsafe fn imageByApplyingGaussianBlurWithSigma( &self, sigma: c_double, ) -> Retained<CIImage>

Create an image by applying a gaussian blur to the receiver.

Parameters:

sigma: The sigma of the gaussian blur to apply to the receiver. If the sigma is very small (less than 0.16) then the receiver is returned.

Returns: An autoreleased CIImage instance or the received image.

pub unsafe fn imageBySettingProperties( &self, properties: &NSDictionary, ) -> Retained<CIImage>

Return a new image by changing the receiver’s metadata properties.

When you create an image, Core Image sets an image’s properties to a metadata dictionary as described here: properties. Use this method to override an image’s metadata properties with new values.

Parameters:

properties: A dictionary of metadata properties akin to the CGImageSourceCopyPropertiesAtIndex() function.

Returns: An autoreleased CIImage instance with a copy of the new properties.

Safety

properties generic should be of the correct type.

pub unsafe fn imageBySamplingLinear(&self) -> Retained<CIImage>

Create an image by changing the receiver’s sample mode to bilinear interpolation.

Returns: An autoreleased CIImage instance with a bilinear sampling.

pub unsafe fn imageBySamplingNearest(&self) -> Retained<CIImage>

Create an image by changing the receiver’s sample mode to nearest neighbor.

Returns: An autoreleased CIImage instance with a nearest sampling.

pub unsafe fn imageByInsertingIntermediate(&self) -> Retained<CIImage>

Create an image that inserts a intermediate that is cacheable

This intermediate will be not be cached if kCIContextCacheIntermediates is false.

Returns: An autoreleased CIImage.

pub unsafe fn imageByInsertingIntermediate_( &self, cache: bool, ) -> Retained<CIImage>

Create an image that inserts a intermediate that is cacheable.

Parameters:

cache: Controls if Core Image caches the returned image.

YES : This intermediate will be cacheable even if kCIContextCacheIntermediates is false.

NO : the intermediate will be not be cached if kCIContextCacheIntermediates is false.

Returns: An autoreleased CIImage.

pub unsafe fn imageByInsertingTiledIntermediate(&self) -> Retained<CIImage>

Create an image that inserts a intermediate that is cached in tiles

This intermediate will be cacheable even if kCIContextCacheIntermediates is false.

Returns: An autoreleased CIImage.

pub unsafe fn imageByApplyingGainMap( &self, gainmap: &CIImage, ) -> Retained<CIImage>

Create an image that applies a gain map Core Image image to the received Core Image image.

The gain map image can be obtained by creating a CIImage instance from NSURL/NSData and setting the kCIImageAuxiliaryHDRGainMap option set to true``.

If the gain map CIImage instance doesn’t have the needed properties metadata, the received image will be returned as-is.

Returns: An autoreleased CIImage instance or the received image.

pub unsafe fn imageByApplyingGainMap_headroom( &self, gainmap: &CIImage, headroom: c_float, ) -> Retained<CIImage>

Create an image that applies a gain map Core Image image with a specified headroom to the received Core Image image.

Parameters:

gainmap: The gain map CIImage instance to apply to the receiver.

headroom: a float value that specify how much headroom the resulting image should have. The headroom value will be limited to between 1.0 (i.e. SDR) and the full headroom allowed by the gain map.

Returns: An autoreleased CIImage instance or the received image.

pub unsafe fn imageBySettingContentHeadroom( &self, headroom: c_float, ) -> Retained<CIImage>

Create an image by changing the receiver’s contentHeadroom property.

Changing this value will alter the behavior of the CIToneMapHeadroom and CISystemToneMap filters.

If the value is set to 0.0 then the returned image’s headroom is unknown.

If the value is set to 1.0 then the returned image is SDR.

If the value is set to greater 1.0 then the returned image is HDR.

Otherwise the returned image’s headroom is unknown.

Returns: An autoreleased CIImage.

pub unsafe fn imageBySettingContentAverageLightLevel( &self, average: c_float, ) -> Retained<CIImage>

Create an image by changing the receiver’s contentAverageLightLevel property.

Changing this value will alter the behavior of the CIToneMapHeadroom and CISystemToneMap filters.

If the value is set to 0.0 or less then the returned image’s contentAverageLightLevel is unknown.

Returns: An autoreleased CIImage.

pub unsafe fn isOpaque(&self) -> bool

Returns YES if the image is known to have and alpha value of 1.0 over the entire image extent.

This property is not atomic.

Safety

This might not be thread-safe.

pub unsafe fn properties(&self) -> Retained<NSDictionary<NSString, AnyObject>>

Returns the metadata properties dictionary of the image.

If the CIImage was created from NSURL or NSData then this dictionary is determined by calling CGImageSourceCopyPropertiesAtIndex().

If the CIImage was created with the kCIImageProperties option, then that dictionary is returned.

If the CIImage was created by applying CIFilter-class or CIKernel then the properties of the root inputImage will be returned.

pub unsafe fn url(&self) -> Option<Retained<NSURL>>

pub unsafe fn contentHeadroom(&self) -> c_float

Returns the content headroom of the image.

If the image headroom is unknown, then the value 0.0 will be returned.

If the image headroom is known, then a value greater than or equal to 1.0 will be returned. A value of 1.0 will be returned if the image is SDR. A value greater than 1.0 will be returned if the image is HDR.

The image headroom may known when a CIImage is first initialized. If the a CIImage is initialized using:

NSURL or NSData : the headroom may be determined by associated metadata or deduced from pixel format or colorSpace information.

CGImage : headroom may be determined by CGImageGetHeadroomInfo() or deduced from pixel format or colorSpace information.

IOSurface : then the headroom will be determined by kIOSurfaceContentHeadroom. or deduced from pixel format or colorSpace information.

CVPixelBuffer : then the headroom will be determined by kCVImageBufferContentLightLevelInfoKey. or deduced from pixel format or colorSpace information.

BitmapData : headroom may be deduced from pixel format or colorSpace information.

If the image is the result of applying a CIFilter-class or CIKernel, this method will return 0.0.

There are exceptions to this. Applying a CIWarpKernel`` or certain ``CIFilter-class`` (e.g. CIGaussianBlur, CILanczosScaleTransform, CIAreaAverageand some others) to an image will result in a ``CIImage`` instance with the samecontentHeadroom` property value.

This property is not atomic.

Safety

This might not be thread-safe.

pub unsafe fn contentAverageLightLevel(&self) -> c_float

Returns the content average light level of the image.

If the image average light level is unknown, then the value 0.0 will be returned.

If the image headroom is known, then a value greater than or equal to 0.0 will be returned.

The image average light level may known when a CIImage is first initialized. If the a CIImage is initialized with a:

CGImage : then the headroom will be determined by CGImageGetContentAverageLightLevel().

CVPixelBuffer : then the headroom will be determined by kCVImageBufferContentLightLevelInfoKey.

If the image is the result of applying a CIFilter-class or CIKernel, this property will return 0.0.

There are exceptions to this. Applying a CIWarpKernel or certain CIFilter-class (e.g. CIGaussianBlur, CILanczosScaleTransform, CIAreaAverage and some others) to an image will result in a CIImage instance with the same contentAverageLightLevel property value.

This property is not atomic.

Safety

This might not be thread-safe.

impl CIImage

Methods declared on superclass NSObject.

pub unsafe fn init(this: Allocated<Self>) -> Retained<Self>

pub unsafe fn new() -> Retained<Self>

impl CIImage

This impl block contains no items.

AutoAdjustment.

impl CIImage

LabConversion.

pub unsafe fn imageByConvertingWorkingSpaceToLab(&self) -> Retained<CIImage>

pub unsafe fn imageByConvertingLabToWorkingSpace(&self) -> Retained<CIImage>